+++ telicent-graph-schema/src/main/java/io/telicent/jena/graphql/fetchers/telicent/graph/StateRelationshipsFetcher.java
@@ -15,14 +15,12 @@ package io.telicent.jena.graphql.fetchers.telicent.graph;
 import graphql.com.google.common.collect.Streams;
 import graphql.schema.DataFetcher;
 import graphql.schema.DataFetchingEnvironment;
-import io.telicent.jena.graphql.execution.telicent.graph.TelicentExecutionContext;
 import io.telicent.jena.graphql.schemas.telicent.graph.models.TelicentGraphNode;
 import io.telicent.jena.graphql.schemas.telicent.graph.models.NonDirectionalRelationship;
 import io.telicent.jena.graphql.schemas.telicent.graph.models.State;
 import org.apache.jena.graph.Node;
 import org.apache.jena.sparql.core.DatasetGraph;
 import org.apache.jena.sparql.core.Quad;
-import org.apache.jena.system.Txn;
 import org.apache.jena.vocabulary.RDF;
 
 import java.util.List;
@@ -32,7 +30,8 @@ import java.util.stream.Stream;
 /**
  * A GraphQL {@link DataFetcher} that finds the relationships for a state
  */
-public class StateRelationshipsFetcher implements DataFetcher<List<NonDirectionalRelationship>> {
+public class StateRelationshipsFetcher
+        extends AbstractLimitOffsetPagingFetcher<State, Quad, List<NonDirectionalRelationship>> {
 
     /**
      * Creates a new fetcher that finds the relationships involving states
@@ -42,39 +41,32 @@ public class StateRelationshipsFetcher implements DataFetcher<List<NonDirectiona
     }
 
     @Override
-    public List<NonDirectionalRelationship> get(DataFetchingEnvironment environment) {
-        TelicentExecutionContext context = environment.getLocalContext();
-        DatasetGraph dsg = context.getDatasetGraph();
-        State target = environment.getSource();
-
-        return Txn.calculateRead(dsg, () -> Streams.concat(outboundRelationships(dsg, target),
-                                                           inboundRelationships(dsg, target))
-                                                   .collect(Collectors.toList()));
+    protected Stream<Quad> select(DataFetchingEnvironment environment, DatasetGraph dsg, State state) {
+        return Streams.concat(outbound(dsg, state), inbound(dsg, state));
     }
 
-    private static Stream<NonDirectionalRelationship> inboundRelationships(DatasetGraph dsg, State target) {
-        return inbound(dsg, target).filter(q -> q.getSubject().isURI() || q.getSubject().isBlank())
-                                   .map(q -> new NonDirectionalRelationship(
-                                           new TelicentGraphNode(q.getPredicate(), dsg.prefixes()),
-                                           new TelicentGraphNode(q.getSubject(), dsg.prefixes())));
-    }
-
-    private static Stream<NonDirectionalRelationship> outboundRelationships(DatasetGraph dsg, State target) {
-        return outbound(dsg, target).filter(q -> q.getObject().isURI() || q.getObject().isBlank())
-                                    .map(q -> new NonDirectionalRelationship(
-                                            new TelicentGraphNode(q.getPredicate(), dsg.prefixes()),
-                                            new TelicentGraphNode(q.getObject(), dsg.prefixes())));
+    @Override
+    protected List<NonDirectionalRelationship> map(DataFetchingEnvironment environment, DatasetGraph dsg, State source,
+                                                   Stream<Quad> input) {
+        return input.map(q -> new NonDirectionalRelationship(new TelicentGraphNode(q.getPredicate(), dsg.prefixes()),
+                                                             new TelicentGraphNode(
+                                                                     source.getStateNode().equals(q.getSubject()) ?
+                                                                     q.getObject() : q.getSubject(), dsg.prefixes())))
+                    .collect(Collectors.toList());
     }
 
     private static Stream<Quad> outbound(DatasetGraph dsg, State target) {
         return dsg.stream(Node.ANY, target.getStateNode(), Node.ANY, Node.ANY)
+                  .filter(q -> q.getObject().isURI() || q.getObject().isBlank())
                   .filter(q -> !q.getPredicate().equals(RDF.type.asNode()) && !q.getObject()
                                                                                 .equals(target.getEntityNode()));
     }
 
     private static Stream<Quad> inbound(DatasetGraph dsg, State target) {
         return dsg.stream(Node.ANY, Node.ANY, Node.ANY, target.getStateNode())
+                  .filter(q -> q.getSubject().isURI() || q.getSubject().isBlank())
                   .filter(q -> dsg.contains(Node.ANY, q.getSubject(), RDF.type.asNode(), Node.ANY));
+
     }
 
 }